Start repository correctly.
I originally had this in my main profile but I am learning how, on this site, to organize by projects into different repositories. I see now how this is super helpful. The next goal will be to get this program working from any computer using githubs tools. I'm sure there is a way to run this as a cli program with out downloading the code to my local computer.